services/grader/requirements.txt
# Python Grader Requirements
# Python 3.11+

# Core dependencies
supabase==2.8.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9

# Image processing
Pillow==10.1.0
numpy==1.26.2
opencv-python-headless==4.8.1.78
scikit-image==0.22.0

# Utilities
tenacity==8.2.3
requests==2.31.0

# Logging
structlog==24.1.0

